move genesis root to chain-config
