+++ fuel-chain-config/Cargo.toml
@@ -12,6 +12,7 @@ description = "Fuel Chain config types"
 
 [dependencies]
 anyhow = "1.0"
+bincode = "1.3"
 fuel-core-interfaces = { path = "../fuel-core-interfaces", version = "0.14.1", features = [
     "serde",
 ] }
