+++ fuel-chain-config/src/config/chain.rs
@@ -4,6 +4,10 @@ use fuel_core_interfaces::common::{
         Address,
         AssetId,
     },
+    prelude::{
+        Hasher,
+        MerkleRoot,
+    },
 };
 use itertools::Itertools;
 use rand::{
@@ -21,6 +25,8 @@ use std::{
     path::PathBuf,
     str::FromStr,
 };
+
+use crate::GenesisCommitment;
 
 use super::{
     coin::CoinConfig,
@@ -120,6 +126,31 @@ impl FromStr for ChainConfig {
                 })
             }
         }
+    }
+}
+
+impl GenesisCommitment for ChainConfig {
+    fn root(&mut self) -> anyhow::Result<MerkleRoot> {
+        // TODO: Hash settlement configuration, consensus block production
+        let config_hash = *Hasher::default()
+            .chain(self.block_gas_limit.to_be_bytes())
+            .chain(self.transaction_parameters.root()?)
+            .chain(self.chain_name.as_bytes())
+            .chain(bincode::serialize(&self.initial_state)?)
+            .finalize();
+
+        Ok(config_hash)
+    }
+}
+
+impl GenesisCommitment for ConsensusParameters {
+    fn root(&mut self) -> anyhow::Result<MerkleRoot> {
+        // TODO: Define hash algorithm for `ConsensusParameters`
+        let params_hash = Hasher::default()
+            .chain(bincode::serialize(&self)?)
+            .finalize();
+
+        Ok(params_hash.into())
     }
 }
