+++ fuel-chain-config/src/config/coin.rs
@@ -1,14 +1,26 @@
-use crate::serialization::{
-    HexNumber,
-    HexType,
+use crate::{
+    serialization::{
+        HexNumber,
+        HexType,
+    },
+    GenesisCommitment,
 };
 use fuel_core_interfaces::{
-    common::fuel_types::{
-        Address,
-        AssetId,
-        Bytes32,
+    common::{
+        fuel_types::{
+            Address,
+            AssetId,
+            Bytes32,
+        },
+        prelude::{
+            Hasher,
+            MerkleRoot,
+        },
     },
-    model::BlockHeight,
+    model::{
+        BlockHeight,
+        Coin,
+    },
 };
 
 use serde::{
@@ -44,4 +56,19 @@ pub struct CoinConfig {
     pub amount: u64,
     #[serde_as(as = "HexType")]
     pub asset_id: AssetId,
+}
+
+impl GenesisCommitment for Coin {
+    fn root(&mut self) -> anyhow::Result<MerkleRoot> {
+        let coin_hash = *Hasher::default()
+            .chain(self.owner)
+            .chain(self.amount.to_be_bytes())
+            .chain(self.asset_id)
+            .chain((*self.maturity).to_be_bytes())
+            .chain([self.status as u8])
+            .chain((*self.block_created).to_be_bytes())
+            .finalize();
+
+        Ok(coin_hash)
+    }
 }
